move tests to utils tests
